% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calculate_r2}
\alias{calculate_r2}
\title{Calculate R^2 and adjusted R^2 for a block and outcomes.
\code{data} is a character vector of file names, a list of matrices, or a single matrix.
\code{outcomes} is a file name or a matrix.
Returns a data frame with columns \code{r2}, \code{adj_r2}, \code{data}, and \code{outcome}.}
\usage{
calculate_r2(data, outcomes)
}
\description{
Calculate R^2 and adjusted R^2 for a block and outcomes.
\code{data} is a character vector of file names, a list of matrices, or a single matrix.
\code{outcomes} is a file name or a matrix.
Returns a data frame with columns \code{r2}, \code{adj_r2}, \code{data}, and \code{outcome}.
}
