% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calculate_r2_ranges}
\alias{calculate_r2_ranges}
\title{Calculate R^2 and adjusted R^2 for ranges of a data matrix and outcomes.
\code{data} is a string file name or a matrix.
\code{outcomes} is a string file name or a matrix.
\code{ranges} is a matrix with 2 columns, the start and end columns to use (inclusive).
Returns a data frame with columns \code{r2} and \code{adj_r2} corresponding to each outcome for each
range in order.}
\usage{
calculate_r2_ranges(data, outcomes, ranges)
}
\description{
Calculate R^2 and adjusted R^2 for ranges of a data matrix and outcomes.
\code{data} is a string file name or a matrix.
\code{outcomes} is a string file name or a matrix.
\code{ranges} is a matrix with 2 columns, the start and end columns to use (inclusive).
Returns a data frame with columns \code{r2} and \code{adj_r2} corresponding to each outcome for each
range in order.
}
