% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{combine_matrices}
\alias{combine_matrices}
\title{DEPRECATED
Combine matrices into a single matrix by columns.
\code{data} is a character vector of file names or a list of matrices.
\code{out} is a file name to write the combined matrix to.
If \code{out} is \code{NULL}, the combined matrix is returned otherwise \code{NULL}.}
\usage{
combine_matrices(data, out)
}
\description{
DEPRECATED
Combine matrices into a single matrix by columns.
\code{data} is a character vector of file names or a list of matrices.
\code{out} is a file name to write the combined matrix to.
If \code{out} is \code{NULL}, the combined matrix is returned otherwise \code{NULL}.
}
