% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_r2}
\alias{compute_r2}
\title{Compute the R^2 value for given actual and predicted vectors.
\code{actual} is a numeric vector of actual values.
\code{predicted} is a numeric vector of predicted values.}
\usage{
compute_r2(actual, predicted)
}
\description{
Compute the R^2 value for given actual and predicted vectors.
\code{actual} is a numeric vector of actual values.
\code{predicted} is a numeric vector of predicted values.
}
