% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_r2_tjur}
\alias{compute_r2_tjur}
\title{Compute Tjur's R^2 value for given actual and predicted vectors.
\code{actual} is a numeric vector of actual values.
\code{predicted} is a numeric vector of predicted probabilities.
See https://doi.org/10.1198/tast.2009.08210}
\usage{
compute_r2_tjur(actual, predicted)
}
\description{
Compute Tjur's R^2 value for given actual and predicted vectors.
\code{actual} is a numeric vector of actual values.
\code{predicted} is a numeric vector of predicted probabilities.
See https://doi.org/10.1198/tast.2009.08210
}
