% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{convert_files}
\alias{convert_files}
\title{Convert files from one format to another.
\code{from} and \code{to} must be character vectors of the same length.}
\usage{
convert_files(from, to, item_type)
}
\description{
Convert files from one format to another.
\code{from} and \code{to} must be character vectors of the same length.
}
