% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{crossprod}
\alias{crossprod}
\title{Computes the cross product of the matrix. Equivalent to \code{t(data) \%*\% data}.
\code{data} must be a string file name or a matrix.
\code{out} is the name of the file to save to.
If \code{out} is \code{NULL}, the cross product is returned otherwise \code{NULL}.}
\usage{
crossprod(data, out)
}
\description{
Computes the cross product of the matrix. Equivalent to \code{t(data) \%*\% data}.
\code{data} must be a string file name or a matrix.
\code{out} is the name of the file to save to.
If \code{out} is \code{NULL}, the cross product is returned otherwise \code{NULL}.
}
