% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cv_elnet_foldids}
\alias{cv_elnet_foldids}
\title{Perform a cross validated elastic net regression with Firth's penalization between each matrix
in a list and each column in another
matrix.
\code{data} is a list of matrix convertible objects.
\code{outcomes} is a matrix convertible object.
\code{alpha} is the elastic net mixing parameter (0 for ridge, 1 for lasso).
\code{nfolds} is the number of folds to use for cross validation.
\code{foldids} is an optional vector of fold ids to use for cross validation.
Returns a data frame with columns \code{data}, \code{outcome}, \code{slopes}, \code{intercept}, \code{lambda}, \code{r2}, and
\code{mse}.}
\usage{
cv_elnet_foldids(data, outcomes, alpha, nfolds, foldids)
}
\description{
Perform a cross validated elastic net regression with Firth's penalization between each matrix
in a list and each column in another
matrix.
\code{data} is a list of matrix convertible objects.
\code{outcomes} is a matrix convertible object.
\code{alpha} is the elastic net mixing parameter (0 for ridge, 1 for lasso).
\code{nfolds} is the number of folds to use for cross validation.
\code{foldids} is an optional vector of fold ids to use for cross validation.
Returns a data frame with columns \code{data}, \code{outcome}, \code{slopes}, \code{intercept}, \code{lambda}, \code{r2}, and
\code{mse}.
}
