% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dedup}
\alias{dedup}
\title{Deduplicate a matrix by a column. The first occurrence of each value is kept.
\code{data} is a string file name or a matrix.
\code{by} is the column to deduplicate by.
\code{out} is a file name to write the deduplicated matrix to or \code{NULL} to return the deduplicated matrix.}
\usage{
dedup(data, by, out)
}
\description{
Deduplicate a matrix by a column. The first occurrence of each value is kept.
\code{data} is a string file name or a matrix.
\code{by} is the column to deduplicate by.
\code{out} is a file name to write the deduplicated matrix to or \code{NULL} to return the deduplicated matrix.
}
