% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{df_combine}
\alias{df_combine}
\title{Combine a potentially nested list of data frames into a single data frame.
\code{data} is a list of lists of data frames.}
\usage{
df_combine(data)
}
\description{
Combine a potentially nested list of data frames into a single data frame.
\code{data} is a list of lists of data frames.
}
