% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{df_sort_asc}
\alias{df_sort_asc}
\title{Mutably sorts a data frame in ascending order by multiple columns in ascending order.
\code{df} is a data frame.
\code{columns} is a character vector of columns to sort by. The sort columns must be numeric
(integer or double), character, or logical vectors.}
\usage{
df_sort_asc(df, columns)
}
\description{
Mutably sorts a data frame in ascending order by multiple columns in ascending order.
\code{df} is a data frame.
\code{columns} is a character vector of columns to sort by. The sort columns must be numeric
(integer or double), character, or logical vectors.
}
