% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{df_split}
\alias{df_split}
\title{Splits a data frame into multiple data frames by a column. This function will mutably sort the
data frame by the column before splitting.
\code{df} is a data frame.
\code{by} is the column to split by. The column must be a numeric (integer or double) or character
vector.}
\usage{
df_split(df, by)
}
\description{
Splits a data frame into multiple data frames by a column. This function will mutably sort the
data frame by the column before splitting.
\code{df} is a data frame.
\code{by} is the column to split by. The column must be a numeric (integer or double) or character
vector.
}
