% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{df_to_matrix_file}
\alias{df_to_matrix_file}
\title{Convert a data frame to a file.
\code{df} must be a numeric data frame.
\code{out} is the name of the file to save to.}
\usage{
df_to_matrix_file(df, out)
}
\description{
Convert a data frame to a file.
\code{df} must be a numeric data frame.
\code{out} is the name of the file to save to.
}
