% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{disable_predicted}
\alias{disable_predicted}
\title{Disable the calculation of predicted values in \code{calculate_r2}.}
\usage{
disable_predicted()
}
\description{
Disable the calculation of predicted values in \code{calculate_r2}.
}
