% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dont_ignore_core_parallel_errors}
\alias{dont_ignore_core_parallel_errors}
\title{Don't ignore errors in core parallel operations.}
\usage{
dont_ignore_core_parallel_errors()
}
\description{
Don't ignore errors in core parallel operations.
}
