% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{enable_predicted}
\alias{enable_predicted}
\title{Enable the calculation of predicted values in \code{calculate_r2}.}
\usage{
enable_predicted()
}
\description{
Enable the calculation of predicted values in \code{calculate_r2}.
}
