% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{extend_matrices}
\alias{extend_matrices}
\title{Extend matrices into a single matrix by rows.
\code{data} is a character vector of file names or a list of matrices.
\code{out} is a file name to write the combined matrix to.
If \code{out} is \code{NULL}, the combined matrix is returned otherwise \code{NULL}.}
\usage{
extend_matrices(data, out)
}
\description{
Extend matrices into a single matrix by rows.
\code{data} is a character vector of file names or a list of matrices.
\code{out} is a file name to write the combined matrix to.
If \code{out} is \code{NULL}, the combined matrix is returned otherwise \code{NULL}.
}
