% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ignore_core_parallel_errors}
\alias{ignore_core_parallel_errors}
\title{Ignore errors in core parallel operations.}
\usage{
ignore_core_parallel_errors()
}
\description{
Ignore errors in core parallel operations.
}
