% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ld_prune}
\alias{ld_prune}
\title{Perform 100\% plink 1.9 compatible LD pruning on the provided bed file.
\itemize{
\item \code{bed} is path to the bed file.
\item \code{window_size} is the size of the window in base pairs.
\item \code{step_size} is the number of variants to step between windows.
\item \code{threshold} is the R^2 threshold above which variants will be pruned.
This function returns a list object with three fields:
\item \code{pruned}: the number of variants pruned.
\item \code{prune_in}: a vector of variant IDs that were kept.
\item \code{prune_out}: a vector of variant IDs that were pruned.
}}
\usage{
ld_prune(bed, window_size, step_size, threshold)
}
\description{
Perform 100\% plink 1.9 compatible LD pruning on the provided bed file.
\itemize{
\item \code{bed} is path to the bed file.
\item \code{window_size} is the size of the window in base pairs.
\item \code{step_size} is the number of variants to step between windows.
\item \code{threshold} is the R^2 threshold above which variants will be pruned.
This function returns a list object with three fields:
\item \code{pruned}: the number of variants pruned.
\item \code{prune_in}: a vector of variant IDs that were kept.
\item \code{prune_out}: a vector of variant IDs that were pruned.
}
}
