% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load}
\alias{load}
\title{Load a matrix convertable object into R.
\code{obj} is a list of matrix convertable objects.
If a single object is provided, the function will return the matrix directly, otherwise it will return a list of matrices.}
\usage{
load(obj)
}
\description{
Load a matrix convertable object into R.
\code{obj} is a list of matrix convertable objects.
If a single object is provided, the function will return the matrix directly, otherwise it will return a list of matrices.
}
