% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_matrix}
\alias{load_matrix}
\title{DEPRECATED
Load a matrix from a file.
\code{file} is the name of the file to load from.}
\usage{
load_matrix(file)
}
\description{
DEPRECATED
Load a matrix from a file.
\code{file} is the name of the file to load from.
}
