% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{logistic_regression_firth}
\alias{logistic_regression_firth}
\title{Compute a logistic regression (using Firth's penalization) between each matrix in a list and each column in another matrix.
\code{data} is a list of matrix convertible objects.
\code{outcomes} is a matrix convertible object.
If there is only one outcome, all rows with NA values in that outcome will be removed.
Returns a data frame with columns \code{slopes}, \code{intercept}, \code{predicted} (if enabled), \code{r2}o,
\code{adj_r2}, \code{r2_tjur}, \code{data}, \code{outcome}, \code{n}, \code{m}, and \code{coefs}.}
\usage{
logistic_regression_firth(data, outcomes)
}
\description{
Compute a logistic regression (using Firth's penalization) between each matrix in a list and each column in another matrix.
\code{data} is a list of matrix convertible objects.
\code{outcomes} is a matrix convertible object.
If there is only one outcome, all rows with NA values in that outcome will be removed.
Returns a data frame with columns \code{slopes}, \code{intercept}, \code{predicted} (if enabled), \code{r2}o,
\code{adj_r2}, \code{r2_tjur}, \code{data}, \code{outcome}, \code{n}, \code{m}, and \code{coefs}.
}
