% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{map_from_pairs}
\alias{map_from_pairs}
\title{Converts two character vectors to a list with the first vector as the names and the second as the values,
only keeping the first occurrence of each name. Essentially creates a hash map.
\code{names} is a character vector of names.
\code{values} is a list of values.}
\usage{
map_from_pairs(names, values)
}
\description{
Converts two character vectors to a list with the first vector as the names and the second as the values,
only keeping the first occurrence of each name. Essentially creates a hash map.
\code{names} is a character vector of names.
\code{values} is a list of values.
}
