% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{match_rows_dir}
\alias{match_rows_dir}
\title{Recursively matches the rows of a directory of matrices to the values in a vector by a column.
\code{from} is the directory to read from.
\code{to} is the directory to write to.
\code{with} is a numeric vector.
\code{by} is the column to match by.}
\usage{
match_rows_dir(from, to, with, by)
}
\description{
Recursively matches the rows of a directory of matrices to the values in a vector by a column.
\code{from} is the directory to read from.
\code{to} is the directory to write to.
\code{with} is a numeric vector.
\code{by} is the column to match by.
}
