% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mean}
\alias{mean}
\title{Compute the mean of a numeric vector.
\code{x} is a numeric vector.}
\usage{
mean(x)
}
\description{
Compute the mean of a numeric vector.
\code{x} is a numeric vector.
}
