% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{median}
\alias{median}
\title{Compute the median of a numeric vector.
\code{x} is a numeric vector.}
\usage{
median(x)
}
\description{
Compute the median of a numeric vector.
\code{x} is a numeric vector.
}
