% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mul}
\alias{mul}
\title{Multiply two matrices.
\code{a} and \code{b} are string file names or matrices.
\code{out} is the name of the file to save to.
If \code{out} is \code{NULL}, the cross product is returned otherwise \code{NULL}.}
\usage{
mul(a, b, out)
}
\description{
Multiply two matrices.
\code{a} and \code{b} are string file names or matrices.
\code{out} is the name of the file to save to.
If \code{out} is \code{NULL}, the cross product is returned otherwise \code{NULL}.
}
