% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{new_column_from_map}
\alias{new_column_from_map}
\title{Add a new column to a data frame from a list of values, matching by the names of the values.
\code{df} is a data frame.
\code{column} is the column to match.
\code{values} is a list of values (a map like from \code{lmutils::map_from_pairs})
\code{new_column} is the new column to create.}
\usage{
new_column_from_map(df, column, values, new_column)
}
\description{
Add a new column to a data frame from a list of values, matching by the names of the values.
\code{df} is a data frame.
\code{column} is the column to match.
\code{values} is a list of values (a map like from \code{lmutils::map_from_pairs})
\code{new_column} is the new column to create.
}
