% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{new_column_from_map_pairs}
\alias{new_column_from_map_pairs}
\title{Compute a new column for a data frame from two character vectors of names and values, matching by the names.
\code{df} is a data frame.
\code{column} is the column to match.
\code{names} is a character vector of names.
\code{values} is a character vector of values.
\code{new_column} is the new column name.}
\usage{
new_column_from_map_pairs(df, column, names, values, new_column)
}
\description{
Compute a new column for a data frame from two character vectors of names and values, matching by the names.
\code{df} is a data frame.
\code{column} is the column to match.
\code{names} is a character vector of names.
\code{values} is a character vector of values.
\code{new_column} is the new column name.
}
