% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{new_column_from_regex}
\alias{new_column_from_regex}
\title{Compute a new column for a data frame from a regex and an existing column.
\code{df} is a data frame.
\code{column} is the column to match.
\code{regex} is the regex to match. The first capture group is used.
\code{new_column} is the new column name.
This function uses the Rust flavor of regex, see https://docs.rs/regex/latest/regex/#syntax for more /* information */.}
\usage{
new_column_from_regex(df, column, regex, new_column)
}
\description{
Compute a new column for a data frame from a regex and an existing column.
\code{df} is a data frame.
\code{column} is the column to match.
\code{regex} is the regex to match. The first capture group is used.
\code{new_column} is the new column name.
This function uses the Rust flavor of regex, see https://docs.rs/regex/latest/regex/#syntax for more /* information */.
}
