% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{num_cores}
\alias{num_cores}
\title{Get the number of cores available.
This is the number of threads that can be used for parallel operations.}
\usage{
num_cores()
}
\description{
Get the number of cores available.
This is the number of threads that can be used for parallel operations.
}
