% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdata_to_matrix}
\alias{rdata_to_matrix}
\title{Convert an RData file to a matrix file.
\code{data} is the name of the RData file.
\code{out} is the name of the file to save to.}
\usage{
rdata_to_matrix(data, out)
}
\description{
Convert an RData file to a matrix file.
\code{data} is the name of the RData file.
\code{out} is the name of the file to save to.
}
