% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{save}
\alias{save}
\title{Saves a list of matrix convertible objects to a character vector of file names.
\code{from} is a list of matrix convertible objects.
\code{to} is a character vector of file names to write to.}
\usage{
save(from, to)
}
\description{
Saves a list of matrix convertible objects to a character vector of file names.
\code{from} is a list of matrix convertible objects.
\code{to} is a character vector of file names to write to.
}
