% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{save_dir}
\alias{save_dir}
\title{Recursively converts a directory of files to the selected format.
\code{from} is the directory to read from.
\code{to} is the directory to write to or \code{NULL} to write to \code{from}.
\code{file_type} is the file extension to write as.}
\usage{
save_dir(from, to, file_type)
}
\description{
Recursively converts a directory of files to the selected format.
\code{from} is the directory to read from.
\code{to} is the directory to write to or \code{NULL} to write to \code{from}.
\code{file_type} is the file extension to write as.
}
