% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{save_matrix}
\alias{save_matrix}
\title{Save a matrix to a file.
\code{mat} must be a double matrix.
\code{out} is the name of the file to save to.}
\usage{
save_matrix(mat, out)
}
\description{
Save a matrix to a file.
\code{mat} must be a double matrix.
\code{out} is the name of the file to save to.
}
