% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sd}
\alias{sd}
\title{Compute the standard deviation of a numeric vector.
\code{x} is a numeric vector.}
\usage{
sd(x)
}
\description{
Compute the standard deviation of a numeric vector.
\code{x} is a numeric vector.
}
