% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_blocks_at_once}
\alias{set_blocks_at_once}
\title{Set the number of blocks to process at once.
\code{blocks_per_chunk} is the number of blocks to process at once.}
\usage{
set_blocks_at_once(blocks_per_chunk)
}
\description{
Set the number of blocks to process at once.
\code{blocks_per_chunk} is the number of blocks to process at once.
}
