% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_core_parallelism}
\alias{set_core_parallelism}
\title{Sets the core parallelism for lmutils.
This is the number of primary operations to perform at once.
\code{num} is the number of main threads.}
\usage{
set_core_parallelism(num)
}
\description{
Sets the core parallelism for lmutils.
This is the number of primary operations to perform at once.
\code{num} is the number of main threads.
}
