% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_num_main_threads}
\alias{set_num_main_threads}
\title{Set the number of main threads to use.
This is the number of primary operations to perform at once.
\code{num} is the number of main threads.}
\usage{
set_num_main_threads(num)
}
\description{
Set the number of main threads to use.
This is the number of primary operations to perform at once.
\code{num} is the number of main threads.
}
