% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_num_worker_threads}
\alias{set_num_worker_threads}
\title{Set the number of worker threads to use.
This is the number of threads to use for parallel operations.
\code{num} is the number of worker threads.}
\usage{
set_num_worker_threads(num)
}
\description{
Set the number of worker threads to use.
This is the number of threads to use for parallel operations.
\code{num} is the number of worker threads.
}
