% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{standardize}
\alias{standardize}
\title{Standardize a matrix. All NaN values are replaced with the mean of the column and each column is scaled to have a mean of 0 and a standard deviation of 1.
\code{data} is a string file name or a matrix.
\code{out} is a file name to write the normalized matrix to.
If \code{out} is \code{NULL}, the normalized matrix is returned otherwise \code{NULL}.}
\usage{
standardize(data, out)
}
\description{
Standardize a matrix. All NaN values are replaced with the mean of the column and each column is scaled to have a mean of 0 and a standard deviation of 1.
\code{data} is a string file name or a matrix.
\code{out} is a file name to write the normalized matrix to.
If \code{out} is \code{NULL}, the normalized matrix is returned otherwise \code{NULL}.
}
