% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{step_aic}
\alias{step_aic}
\title{Perform stepwise selection of a logistic regression model using AIC.
\code{data} a single matrix convertible object as the design matrix. It must have column names.
\code{outcomes} is a double vector containing the binary outcomes.
\code{from} either 'null' or 'full' to indicate whether to start from a null model or a full model.
\code{direction} is either 'forward', 'backward', or 'both' to indicate the direction of the stepwise selection.
Returns a list object with fields \code{slopes}, \code{intercept}, \code{predicted}, \code{r2}, \code{adj_r2}, \code{aic},
and \code{coefs}.}
\usage{
step_aic(data, outcomes, from, direction)
}
\description{
Perform stepwise selection of a logistic regression model using AIC.
\code{data} a single matrix convertible object as the design matrix. It must have column names.
\code{outcomes} is a double vector containing the binary outcomes.
\code{from} either 'null' or 'full' to indicate whether to start from a null model or a full model.
\code{direction} is either 'forward', 'backward', or 'both' to indicate the direction of the stepwise selection.
Returns a list object with fields \code{slopes}, \code{intercept}, \code{predicted}, \code{r2}, \code{adj_r2}, \code{aic},
and \code{coefs}.
}
