% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{to_matrix}
\alias{to_matrix}
\title{Convert a data frame to a file.
\code{df} must be a numeric data frame, numeric matrix, or a string RData file name.
\code{out} is the name of the file to save to.
If \code{out} is \code{NULL}, the matrix is returned otherwise \code{NULL}.}
\usage{
to_matrix(df, out)
}
\description{
Convert a data frame to a file.
\code{df} must be a numeric data frame, numeric matrix, or a string RData file name.
\code{out} is the name of the file to save to.
If \code{out} is \code{NULL}, the matrix is returned otherwise \code{NULL}.
}
