% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{var}
\alias{var}
\title{Compute the variance of a numeric vector.
\code{x} is a numeric vector.}
\usage{
var(x)
}
\description{
Compute the variance of a numeric vector.
\code{x} is a numeric vector.
}
