% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calculate_r2}
\alias{calculate_r2}
\title{Calculate R^2 and adjusted R^2 for a block and outcomes.
\code{data} is a list of matrix convertible objects.
\code{outcomes} is a matrix convertible object.
Returns a data frame with columns \code{r2}, \code{adj_r2}, \code{data}, \code{outcome}, \code{n}, \code{m}, and \code{predicted}.}
\usage{
calculate_r2(data, outcomes)
}
\description{
Calculate R^2 and adjusted R^2 for a block and outcomes.
\code{data} is a list of matrix convertible objects.
\code{outcomes} is a matrix convertible object.
Returns a data frame with columns \code{r2}, \code{adj_r2}, \code{data}, \code{outcome}, \code{n}, \code{m}, and \code{predicted}.
}
