% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{crossprod}
\alias{crossprod}
\title{Computes the cross product of the matrix. Equivalent to \code{t(data) \%*\% data}.
\code{data} is a list of matrix convertable objects.
\code{out} is a standard output file.}
\usage{
crossprod(data, out)
}
\description{
Computes the cross product of the matrix. Equivalent to \code{t(data) \%*\% data}.
\code{data} is a list of matrix convertable objects.
\code{out} is a standard output file.
}
