% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dedup}
\alias{dedup}
\title{Deduplicate a matrix by a column. The first occurrence of each value is kept.
\code{data} is a list of matrix convertible objects.
\code{by} is the column to deduplicate by.
\code{out} is a standard output file.}
\usage{
dedup(data, by, out)
}
\description{
Deduplicate a matrix by a column. The first occurrence of each value is kept.
\code{data} is a list of matrix convertible objects.
\code{by} is the column to deduplicate by.
\code{out} is a standard output file.
}
