% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{linear_regression}
\alias{linear_regression}
\title{Compute a linear regression between each matrix in a list and a each column in another matrix.
\code{data} is a list of matrix convertable objects.
\code{outcomes} is a matrix convertable object.
Returns a data frame with columns \code{slopes}, \code{intercept}, \code{predicted} (if enabled), \code{r2},
\code{adj_r2}, \code{data}, \code{outcome}, \code{n}, and \code{m}.}
\usage{
linear_regression(data, outcomes)
}
\description{
Compute a linear regression between each matrix in a list and a each column in another matrix.
\code{data} is a list of matrix convertable objects.
\code{outcomes} is a matrix convertable object.
Returns a data frame with columns \code{slopes}, \code{intercept}, \code{predicted} (if enabled), \code{r2},
\code{adj_r2}, \code{data}, \code{outcome}, \code{n}, and \code{m}.
}
