% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mul}
\alias{mul}
\title{Multiply two matrices. Equivalent to \code{a \%*\% b}.
\code{a} is a list of matrix convertable objects.
\code{b} is a list of matrix convertable objects.
\code{out} is a standard non-mutating output.}
\usage{
mul(a, b, out)
}
\description{
Multiply two matrices. Equivalent to \code{a \%*\% b}.
\code{a} is a list of matrix convertable objects.
\code{b} is a list of matrix convertable objects.
\code{out} is a standard non-mutating output.
}
