% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{new_column_from_map}
\alias{new_column_from_map}
\title{Compute a new column for a data frame from a list of values and an existing column, matching by the names of the values.
\code{column} is the column to match.
\code{values} is a named list of values.
\code{new_column} is the new column name.}
\usage{
new_column_from_map(df, column, values, new_column)
}
\description{
Compute a new column for a data frame from a list of values and an existing column, matching by the names of the values.
\code{column} is the column to match.
\code{values} is a named list of values.
\code{new_column} is the new column name.
}
