% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove rows from a matrix.
\code{data} is a matrix.
\code{rows} is a vector of row indices to remove.
\code{out} is a file name to write the matrix with the rows removed to.
If \code{out} is \code{NULL}, the matrix with the rows removed is returned otherwise \code{NULL}.}
\usage{
remove_rows(data, rows, out)
}
\description{
Remove rows from a matrix.
\code{data} is a matrix.
\code{rows} is a vector of row indices to remove.
\code{out} is a file name to write the matrix with the rows removed to.
If \code{out} is \code{NULL}, the matrix with the rows removed is returned otherwise \code{NULL}.
}
