% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove rows from a matrix.
\code{data} is a list of matrix convertable objects.
\code{rows} is a vector of row indices to remove (1-based).
\code{out} is a standard non-mutating output.}
\usage{
remove_rows(data, rows, out)
}
\description{
Remove rows from a matrix.
\code{data} is a list of matrix convertable objects.
\code{rows} is a vector of row indices to remove (1-based).
\code{out} is a standard non-mutating output.
}
